requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
djangorestframework==3.15.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
fsspec==2024.2.0
greenlet==3.0.3
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.2.1
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
Pillow==10.1.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
sympy==1.12
tbb==2021.11.0
torch==2.4.0.dev20240520+cu124
torchaudio==2.2.0.dev20240520+cu124
torchvision==0.19.0.dev20240520+cu124
typing_extensions==4.8.0
tzdata==2024.1
Werkzeug==3.0.3
